refactor symb module
