+++ include/math/symb/Equation.cpp
@@ -1,0 +1,2 @@
+#include "Equation.h"
+
